Try splitting reducers
