+++ app/app.js
@@ -1,25 +1,25 @@
 import 'whatwg-fetch';
 import ReactDOM from 'react-dom';
 import React from 'react';
-import { createStore, applyMiddleware } from 'redux';
-import { Provider } from 'react-redux';
+import {createStore, applyMiddleware} from 'redux';
+import {Provider} from 'react-redux';
 import thunk from 'redux-thunk';
 
-import reducers from './reducers';
+import combineReducers from './store/reducers';
 import App from './containers/App';
 
 const middlewares = [thunk];
 
 const store = createStore(
-    reducers,
-    applyMiddleware(...middlewares)
+  combineReducers,
+  applyMiddleware(...middlewares)
 );
 
-const Root = ({ store }) => (
-    <Provider store={store}>
-        <App />
-    </Provider>
+const Root = ({store}) => (
+  <Provider store={store}>
+    <App />
+  </Provider>
 );
 
 
-ReactDOM.render(<Root store={store} />, document.getElementById('app-root'));
+ReactDOM.render(<Root store={store}/>, document.getElementById('app-root'));
